// src/app/models/Film.ts
export class Film {
    title: string;
    // tslint:disable-next-line:variable-name
    opening_crawl: string;
    director: string;
    producer: string;
    // tslint:disable-next-line:variable-name
    release_date: string;
    characters: string[];
    planets: string[];
    starships: string[];
    vehicles: string[];
    species: string[];
    created: string;
    edited: string;
    url: string;
}